// ui/src/generated/RuntimeLaunchBinding.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type RuntimeLaunchBinding = { r_executable: string, ark_executable: string, environment_realization_id: string | null, library_path: string | null, checkpoint_helper_path: string | null, };